refactor: remove raw avatar ulr
