+++ src/helpers/members/avatar_url.ts
@@ -1,17 +1,15 @@
-import { Member } from "../../structures/mod.ts";
-import { rawAvatarURL } from "./raw_avatar_url.ts";
+import { endpoints } from "../../util/constants.ts";
+import { formatImageURL } from "../../util/utils.ts";
 
-/** The users custom avatar or the default avatar */
+/** The users custom avatar or the default avatar if you don't have a member object. */
 export function avatarURL(
-  member: Member,
+  userId: string,
+  discriminator: string,
+  avatar?: string | null,
   size: ImageSize = 128,
   format?: ImageFormats,
 ) {
-  return rawAvatarURL(
-    member.id,
-    member.discriminator,
-    member.avatar,
-    size,
-    format,
-  );
+  return avatar
+    ? formatImageURL(endpoints.USER_AVATAR(userId, avatar), size, format)
+    : endpoints.USER_DEFAULT_AVATAR(Number(discriminator) % 5);
 }
